+++ packages/extract-emails/test/extract-emails.spec.ts
@@ -93,10 +93,10 @@ describe('ExtractEmails', function () {
       },
     },
     {
-      version: 'v0.0.3',
+      version: 'v0.0.4',
       type: SupportedSchemas.COA,
-      certificatePath: `${__dirname}/../../../fixtures/CoA/v0.0.3/valid_cert.json`,
-      certificate: require('../../../fixtures/CoA/v0.0.3/valid_cert.json'),
+      certificatePath: `${__dirname}/../../../fixtures/CoA/v0.0.4/valid_cert.json`,
+      certificate: require('../../../fixtures/CoA/v0.0.4/valid_cert.json'),
       expectedSenders: {
         [SenderRoles.Manufacturer]: {
           emails: [certificateEmail],